+++ CHANGES.md
@@ -5,7 +5,8 @@ CHANGELOG
 0.3.50.dev0     (XXXX-XX-XX)
 ----------------------------
 
-* 
+* Allow to define which property should be included in vector tile
+* Fix cases when layer has not yet schema definition
 
 
 0.3.49          (2020-10-16)